Patch from Yichao Yu
Choqok's TextEdit uses the minimum length of the scrollbar as it's
height (through QAbstractScrollArea::minimumSizeHint) and it also sets a
hard limit on the height of the editor. These cause two problem:
1. The height of the TextEdit is theme dependent and for a theme that
allows short scrollbar (which is perfectly legal), the text box can be
too small.
2. When the use specifies a larger font (e.g. 32-72) the TextEditor will
not expand along with it leaving only one (or less) visible lines.
REVIEW:114022