Disconnect destroyed() signal from all tab widgets in ChoqokTabBar dtor
Because the QWidget dtor will delete its children first, it is the
responsability of any subclass to disconnect any signal that might be
emitted as a result of a child's destruction.  Otherwise, that signal
could end up attempting to invoke a method on the now-destroyed parent.

The ChoqokTabBar dtor must therefore disconnect the destroyed() signal
from all children of its own QStackedWidget.

BUG: 400117