Disconnect self-connected signals emitted from a parent class
Signals connected from an object to itself should be disconnected in the
dtor if said signal is emitted from the parent class.  Otherwise, if
that signal ends up being emitted in the parent dtor, it will attempt to
invoke a method on the now-destroyed child object.

(In all fairness, only one of these connections, namely the one to
TextEdit::textChanged(), is actually problematic, as all others fall
into the "you can no longer click on stuff at this point anyway"
category.  Still, better safe than sorry.)

BUG: 455644