Merge two opposite `if` statements in ChoqokTabBar::moveTab()
g++ is apparently not smart enough to realize that one of these two `if`
statements will be executed no matter what, and throws a
`maybe-uninitialized` warning as a result.

That said, with `from == to` being handled above, this is clearly the
equivalent of `if (A)` followed by `if (!A)`, so it makes sense to merge
the both of them.